refactor (components rendering): Код отрисовки компонентов разделён на модули
