+++ src/components/films-section.js
@@ -1,0 +1,8 @@
+const createFilmsSectionTemplate = () => {
+  return (
+    `<section class="films">
+    </section>`
+  );
+};
+
+export {createFilmsSectionTemplate};
