+++ src/components/top-rated-films.js
@@ -1,0 +1,10 @@
+const createTopRatedFilmsTemplate = () => {
+  return (
+    `<section class="films-list--extra">
+      <h2 class="films-list__title">Top rated</h2>
+      <div class="films-list__container" id="top-rated-films-list"></div>
+    </section>`
+  );
+};
+
+export {createTopRatedFilmsTemplate};
